feat: add worked broadcast

+++ app/database/models/users.py
@@ -1,4 +1,4 @@
-from sqlalchemy import Column, String, Integer, Boolean, ForeignKey, Text, TIMESTAMP, DateTime
+from sqlalchemy import Column, String, Integer, Boolean, ForeignKey, Text, TIMESTAMP, DateTime, BigInteger
 from sqlalchemy.dialects.postgresql import UUID
 from sqlalchemy.orm import relationship, Mapped, mapped_column
 from app.database.psql import Base
@@ -10,7 +10,7 @@ from datetime import datetime
 class User(Base):
     __tablename__ = 'users'
 
-    telegram_id: Mapped[int] = mapped_column(Integer, unique=True,primary_key=True, nullable=False)
+    telegram_id: Mapped[int] = mapped_column(BigInteger, unique=True,primary_key=True, nullable=False)
     access_token : Mapped[str] = mapped_column(String, nullable=True)
     refresh_token : Mapped[str] = mapped_column(String, nullable=True)
     group_id: Mapped[int] = mapped_column(Integer, ForeignKey('groups.id', ondelete='CASCADE'), nullable=True)
